remove old combox, rename ComboBox2 to ComboBox
