+++ resources/js/components/MemberList.vue
@@ -113,7 +113,7 @@
 <script lang="ts">
 import GroupTitle from "./GroupTitle.vue";
 import { dayjs, $can } from "@/utils";
-import ComboBox from "./ComboBox2.vue";
+import ComboBox from "./ComboBox.vue";
 
 export default {
   components: {
